Extract story from test
